Update to 1.20.4
- Switch to use NeoForge
- Update to Arch Loom 1.5
- Update to gradle 8.4
- Use builtin Mixin Extras for both fabric and neo
- Adjust ModifyExpressionValue for setInWater and setEyeInFluid with such now existing in ServerPlayerMixin
- Switch to V2 WrapWithCondition
