+++ common/build.gradle
@@ -6,7 +6,7 @@ dependencies {
     // We depend on fabric loader here to use the fabric @Environment annotations and get the mixin dependencies
     // Do NOT use other classes from fabric loader
     modImplementation "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
-    implementation annotationProcessor("com.github.LlamaLad7:MixinExtras:${rootProject.mixin_extras_version}")
+    implementation annotationProcessor("io.github.llamalad7:mixinextras-common:${rootProject.mixin_extras_version}")
 }
 
 publishing {
