--- forge/build.gradle
@@ -1,114 +1,0 @@
-plugins {
-    id "com.github.johnrengelman.shadow" version "7.1.2"
-    id "com.matthewprenger.cursegradle"
-}
-
-architectury {
-    platformSetupLoomIde()
-    forge()
-}
-
-loom {
-    forge {
-        mixinConfig "expandability.mixins.json"
-        mixinConfig "expandability-common.mixins.json"
-    }
-}
-
-configurations {
-    common
-    shadowCommon // Don't use shadow from the shadow plugin because we don't want IDEA to index this.
-    compileClasspath.extendsFrom common
-    runtimeClasspath.extendsFrom common
-    developmentForge.extendsFrom common
-}
-
-dependencies {
-    forge "net.minecraftforge:forge:${rootProject.minecraft_version}-${rootProject.forge_version}"
-    implementation include(forgeRuntimeLibrary(annotationProcessor("com.github.LlamaLad7:MixinExtras:${rootProject.mixin_extras_version}")))
-
-    common(project(path: ":common", configuration: "namedElements")) { transitive false }
-    shadowCommon(project(path: ":common", configuration: "transformProductionForge")) { transitive false }
-}
-
-processResources {
-    inputs.property "version", project.version
-
-    filesMatching("META-INF/mods.toml") {
-        expand "version": project.version
-    }
-}
-
-shadowJar {
-    exclude "fabric.mod.json"
-
-    configurations = [project.configurations.shadowCommon]
-    archiveClassifier = "dev-shadow"
-}
-
-remapJar {
-    input.set shadowJar.archiveFile
-    dependsOn shadowJar
-    archiveClassifier = null
-}
-
-jar {
-    archiveClassifier = "dev"
-}
-
-sourcesJar {
-    def commonSources = project(":common").sourcesJar
-    dependsOn commonSources
-    from commonSources.archiveFile.map { zipTree(it) }
-}
-
-components.java {
-    withVariantsFromConfiguration(project.configurations.shadowRuntimeElements) {
-        skip()
-    }
-}
-
-publishing {
-    publications {
-        mavenForge(MavenPublication) {
-            artifactId = rootProject.archives_base_name + "-" + project.name
-            from components.java
-        }
-    }
-
-    repositories {
-        maven {
-            url = "https://maven.florens.be/releases"
-            credentials {
-                username = "florens"
-                password = System.getenv("MAVEN_PASS")
-            }
-        }
-    }
-}
-
-curseforge {
-    if (System.getenv("CURSE_API_KEY") != null) {
-        apiKey = System.getenv("CURSE_API_KEY")
-
-        project {
-            id = project.curse_project
-            releaseType = "release"
-            addGameVersion "1.20"
-            addGameVersion "1.20.1"
-            addGameVersion "Forge"
-            mainArtifact(remapJar.archivePath) {
-                displayName = "[Forge $rootProject.supported_versions] v$project.version"
-            }
-            afterEvaluate {
-                uploadTask.dependsOn("build")
-            }
-        }
-    }
-    options {
-        forgeGradleIntegration = false
-        javaVersionAutoDetect = false
-    }
-}
-
-rootProject.tasks.getByName("curseforgePublish").dependsOn tasks.getByName("curseforge")
