--- forge/src/main/java/be/florens/expandability/forge/ExpandAbilityForge.java
@@ -1,12 +1,0 @@
-package be.florens.expandability.forge;
-
-import be.florens.expandability.ExpandAbility;
-import net.minecraftforge.fml.common.Mod;
-
-@Mod(ExpandAbility.MOD_ID)
-public class ExpandAbilityForge {
-
-	public ExpandAbilityForge() {
-		ExpandAbility.init();
-	}
-}
