--- forge/src/main/resources/META-INF/mods.toml
@@ -1,28 +1,0 @@
-modLoader = "javafml"
-loaderVersion = "[46,)"
-issueTrackerURL = "https://github.com/florensie/expandability/issues"
-license = "MIT"
-
-[[mods]]
-modId = "expandability"
-version = "${version}"
-displayName = "ExpandAbility"
-authors = "Florens"
-description = '''
-Library mod that provides increased control over vanilla abilities
-'''
-logoFile = "icon.png"
-
-[[dependencies.expandability]]
-modId = "forge"
-mandatory = true
-versionRange = "[46,)"
-ordering = "NONE"
-side = "BOTH"
-
-[[dependencies.expandability]]
-modId = "minecraft"
-mandatory = true
-versionRange = "[1.20,)"
-ordering = "NONE"
-side = "BOTH"
