+++ gradle.properties
@@ -3,13 +3,13 @@
 org.gradle.jvmargs=-Xmx4G
 
 # General properties
-supported_versions=1.20-1.20.1
+supported_versions=1.20.4
 curse_project=465066
-minecraft_version=1.20
+minecraft_version=1.20.4
 
 # Allow temporarily disabling certain platforms
 # For example to build and release a Fabric version when Forge is not yet available
-enabled_platforms=fabric,forge
+enabled_platforms=fabric,neoforge
 
 # Mod properties
 mod_version=9.0.0
@@ -17,11 +17,11 @@ maven_group=be.florens
 archives_base_name=expandability
 
 # Common dependencies
-mixin_extras_version=0.1.1
+mixin_extras_version=0.3.5
 
 # Fabric dependencies
-fabric_loader_version=0.14.21
-fabric_api_version=0.83.0+1.20
+fabric_loader_version=0.15.9
+fabric_api_version=0.96.11+1.20.4
 
-# Forge dependencies
-forge_version=46.0.14
+# NeoForge dependencies
+neoforge_version=20.4.159-beta
