+++ neoforge/build.gradle
@@ -5,7 +5,14 @@ plugins {
 
 architectury {
     platformSetupLoomIde()
-    fabric()
+    neoForge()
+}
+
+loom {
+//    forge {
+//        mixinConfig "expandability.mixins.json"
+//        mixinConfig "expandability-common.mixins.json"
+//    }
 }
 
 configurations {
@@ -13,26 +20,31 @@ configurations {
     shadowCommon // Don't use shadow from the shadow plugin because we don't want IDEA to index this.
     compileClasspath.extendsFrom common
     runtimeClasspath.extendsFrom common
-    developmentFabric.extendsFrom common
+    developmentNeoForge.extendsFrom common
+}
+
+repositories {
+    maven { url "https://maven.neoforged.net/releases/" }
 }
 
 dependencies {
-    modImplementation "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
-    modApi "net.fabricmc.fabric-api:fabric-api:${rootProject.fabric_api_version}"
+    neoForge "net.neoforged:neoforge:${rootProject.neoforge_version}"
 
     common(project(path: ":common", configuration: "namedElements")) { transitive false }
-    shadowCommon(project(path: ":common", configuration: "transformProductionFabric")) { transitive false }
+    shadowCommon(project(path: ":common", configuration: "transformProductionNeoForge")) { transitive false }
 }
 
 processResources {
     inputs.property "version", project.version
 
-    filesMatching("fabric.mod.json") {
+    filesMatching("META-INF/mods.toml") {
         expand "version": project.version
     }
 }
 
 shadowJar {
+    exclude "fabric.mod.json"
+
     configurations = [project.configurations.shadowCommon]
     archiveClassifier = "dev-shadow"
 }
@@ -61,7 +73,7 @@ components.java {
 
 publishing {
     publications {
-        mavenFabric(MavenPublication) {
+        mavenForge(MavenPublication) {
             artifactId = rootProject.archives_base_name + "-" + project.name
             from components.java
         }
@@ -87,10 +99,9 @@ curseforge {
             releaseType = "release"
             addGameVersion "1.20"
             addGameVersion "1.20.1"
-            addGameVersion "Fabric"
-            addGameVersion "Quilt"
+            addGameVersion "Forge"
             mainArtifact(remapJar.archivePath) {
-                displayName = "[Fabric $rootProject.supported_versions] v$project.version"
+                displayName = "[Forge $rootProject.supported_versions] v$project.version"
             }
             afterEvaluate {
                 uploadTask.dependsOn("build")
