+++ neoforge/src/main/java/be/florens/expandability/mixin/forge/swimming/LivingEntityMixin.java
@@ -1,0 +1,31 @@
+package be.florens.expandability.mixin.forge.swimming;
+
+import be.florens.expandability.EventDispatcher;
+import be.florens.expandability.Util;
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.player.Player;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(LivingEntity.class)
+public abstract class LivingEntityMixin {
+
+    @ModifyExpressionValue(method = "travel", require = 2, at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;isInFluidType(Lnet/minecraft/world/level/material/FluidState;)Z"))
+    private boolean setInWater(boolean original) {
+        if ((Object) this instanceof Player player) {
+            return Util.processEventResult(EventDispatcher.onPlayerSwim(player), original);
+        }
+
+        return original;
+    }
+
+    @ModifyExpressionValue(method = "aiStep", require = 2, at = @At(value = "INVOKE", target = "Lnet/neoforged/neoforge/fluids/FluidType;isAir()Z"))
+    private boolean setIsAir(boolean original) {
+        if ((Object) this instanceof Player player) {
+            return Util.processEventResult(EventDispatcher.onPlayerSwim(player), false, true, original);
+        }
+
+        return original;
+    }
+}
