+++ neoforge/src/main/java/be/florens/expandability/neoforge/ExpandAbilityForge.java
@@ -1,0 +1,12 @@
+package be.florens.expandability.neoforge;
+
+import be.florens.expandability.ExpandAbility;
+import net.neoforged.fml.common.Mod;
+
+@Mod(ExpandAbility.MOD_ID)
+public class ExpandAbilityForge {
+
+	public ExpandAbilityForge() {
+		ExpandAbility.init();
+	}
+}
