+++ settings.gradle
@@ -3,15 +3,16 @@ pluginManagement {
         maven { url "https://maven.fabricmc.net/" }
         maven { url "https://maven.architectury.dev/" }
         maven { url "https://maven.minecraftforge.net/" }
+        maven { url "https://maven.neoforged.net/releases/" }
         gradlePluginPortal()
     }
 }
 
 include("common")
 include("fabric")
-include("forge")
+include("neoforge")
 include("testmod-common")
 include("testmod-fabric")
-include("testmod-forge")
+include("testmod-neoforge")
 
 rootProject.name = "expandability"
