+++ testmod-fabric/build.gradle
@@ -36,7 +36,6 @@ loom {
 dependencies {
     modImplementation "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
     modImplementation "net.fabricmc.fabric-api:fabric-api:${rootProject.fabric_api_version}"
-    runtimeOnly "com.github.LlamaLad7:MixinExtras:${rootProject.mixin_extras_version}"
 
     implementation project(path: ":fabric", configuration: "namedElements")
     common(project(path: ":common", configuration: "namedElements")) { transitive false }
