--- testmod-forge/build.gradle
@@ -1,36 +1,0 @@
-plugins {
-    id "com.github.johnrengelman.shadow" version "7.1.2"
-}
-
-loom {
-    mods {
-        forge {
-            sourceSet project(":forge").sourceSets.main
-        }
-    }
-
-    forge {
-        mixinConfig "expandability.mixins.json"
-    }
-}
-
-architectury {
-    platformSetupLoomIde()
-    forge()
-}
-
-configurations {
-    common
-    compileClasspath.extendsFrom common
-    runtimeClasspath.extendsFrom common
-    developmentForge.extendsFrom common
-}
-
-dependencies {
-    forge "net.minecraftforge:forge:${gradle.rootProject.architectury.minecraft}-${rootProject.forge_version}"
-    runtimeOnly forgeRuntimeLibrary("com.github.LlamaLad7:MixinExtras:${rootProject.mixin_extras_version}")
-
-    implementation project(path: ":forge", configuration: "namedElements")
-    common(project(path: ":common", configuration: "namedElements")) { transitive false }
-    common(project(path: ":testmod-common", configuration: "namedElements")) { transitive false }
-}
