--- testmod-forge/src/main/resources/META-INF/mods.toml
@@ -1,19 +1,0 @@
-modLoader = "javafml"
-loaderVersion = "[32,)"
-license = "MIT"
-
-[[mods]]
-modId = "expandabilitytest"
-version = "1.0.0"
-displayName = "ExpandAbility Test"
-authors = "Florens"
-description = '''
-Test mod for ExpandAbility
-'''
-
-[[dependencies.expandabilitytest]]
-modId = "expandability"
-mandatory = true
-ordering = "NONE"
-side = "BOTH"
-versionRange = "[0,)" # fuck you froge
