+++ testmod-neoforge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,19 @@
+modLoader = "javafml"
+loaderVersion = "[1,)"
+license = "MIT"
+
+[[mods]]
+modId = "expandabilitytest"
+version = "1.0.0"
+displayName = "ExpandAbility Test"
+authors = "Florens"
+description = '''
+Test mod for ExpandAbility
+'''
+
+[[dependencies.expandabilitytest]]
+modId = "expandability"
+required = true
+ordering = "NONE"
+side = "BOTH"
+versionRange = "[0,)" # fuck you froge
