+++ testmod-neoforge/src/main/resources/pack.mcmeta
